requirements.txt
# packages in environment at /Users/athinaapostolelli/miniforge3/envs/hypnose-analysis:
#
# Name                    Version                   Build  Channel
aeon=1.1.0
annotated-types=0.7.0
appnope=0.1.4
asttokens=3.0.0
bzip2=1.0.8
ca-certificates=2025.6.15
comm=0.2.2
contourpy=1.3.2
cycler=0.12.1
debugpy=1.8.14
decorator=5.2.1
deprecated=1.2.18
dotmap=1.3.30
exceptiongroup=1.3.0
executing=2.2.0
fonttools=4.58.0
harp-python=0.4.1
hypnose=0.1.0
importlib-metadata=8.7.0
ipykernel=6.29.5
ipympl=0.9.7
ipython=9.2.0
ipython_pygments_lexers=1.1.1
ipywidgets=8.1.7
jedi=0.19.2
joblib=1.5.1
jupyter_client=8.6.3
jupyter_core=5.8.1
jupyterlab-widgets=3.0.15
kiwisolver=1.4.8
krb5=1.21.3
libcxx=20.1.7
libedit=3.1.20250104
libexpat=2.7.0
libffi=3.4.6
liblzma=5.8.1
libmpdec=4.0.0
libsodium=1.0.20
libsqlite=3.49.2
libzlib=1.3.1
llvmlite=0.44.0
matplotlib=3.10.3
matplotlib-inline=0.1.7
ncurses=6.5
nest-asyncio=1.6.0
numba=0.61.2
numpy=1.26.4
opencv-python=4.11.0.86
openssl=3.5.0
packaging=25.0
pandas=2.2.3
parso=0.8.4
pexpect=4.9.0
pickleshare=0.7.5
pillow=11.2.1
pip=25.1.1
platformdirs=4.3.8
prompt-toolkit=3.0.51
psutil=7.0.0
ptyprocess=0.7.0
pure_eval=0.2.3
pydantic=2.11.5
pydantic-core=2.33.2
pydantic-yaml=1.5.1
pygments=2.19.1
pyparsing=3.2.3
python=3.13.3
python-dateutil=2.9.0
python_abi=3.13
pytz=2025.2
pyyaml=6.0.2
pyzmq=27.0.0
readline=8.2
ruamel-yaml=0.18.11
ruamel-yaml-clib=0.2.12
scikit-learn=1.6.1
scipy=1.15.3
six=1.17.0
stack_data=0.6.3
swc-aeon=0.1.0
threadpoolctl=3.6.0
tk=8.6.13
tornado=6.5.1
traitlets=5.14.3
typing-extensions=4.13.2
typing-inspection=0.4.1
typing_extensions=4.14.0
tzdata=2025.2
wcwidth=0.2.13
widgetsnbextension=4.0.14
wrapt=1.17.2
zeromq=4.3.5
zipp=3.23.0
